Moved metrics to internal package
