+++ sdk/trace/src/main/java/io/opentelemetry/sdk/trace/export/BatchSpanProcessor.java
@@ -13,8 +13,8 @@ import io.opentelemetry.sdk.trace.ReadWriteSpan;
 import io.opentelemetry.sdk.trace.ReadableSpan;
 import io.opentelemetry.sdk.trace.SpanProcessor;
 import io.opentelemetry.sdk.trace.data.SpanData;
-import io.opentelemetry.sdk.trace.export.metrics.SpanProcessorMetrics;
 import io.opentelemetry.sdk.trace.internal.JcTools;
+import io.opentelemetry.sdk.trace.internal.metrics.SpanProcessorMetrics;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
