+++ sdk/trace/src/main/java/io/opentelemetry/sdk/trace/internal/metrics/SemConvSpanProcessorMetrics.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.trace.export.metrics;
+package io.opentelemetry.sdk.trace.internal.metrics;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
